requirements.txt
streamlit==1.28.1
PyMuPDF==1.23.8
spacy==3.7.2
nltk==3.8.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
wordcloud==1.9.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
fpdf2==2.7.6
python-docx==1.1.0
textstat==0.7.3
transformers==4.35.2
torch==2.1.1
openai==1.3.7
python-dotenv==1.0.0
Pillow==10.1.0